pyproject.toml
[tool.poetry]
name = "opera-lyrics-test"
version = "0.1.0"
description = ""
authors = ["None"]

[tool.poetry.dependencies]
python = "^3.8"
lyricsgenius = "^3.0.1"
python-dotenv = "^0.17.0"
ipykernel = "^5.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
